Update header with tier information
Previously, the header was only updated with the latest tier information
when the user refreshed the page, which meant it was not obvious to the
user what the current state of their payment plan was. Updated the
activation and deactivation logic to update the allowance element in the
header appropriately.

https://trello.com/c/EJLrc4J5